+++ src-tauri/crates/tenant-service/src/products/dto.rs
@@ -1,4 +1,3 @@
-use sea_orm::entity::prelude::Decimal;
 use sea_orm::FromQueryResult;
 use serde::{Deserialize, Serialize};
 use specta::Type;
@@ -19,14 +18,14 @@ pub struct SelectProducts {
     pub created_at: String,
     pub description: Option<String>,
     pub image: Option<String>,
-    #[specta(type = f64)]
-    pub selling_price: Option<Decimal>,
-    #[specta(type = f64)]
-    pub purchase_price: Option<Decimal>,
-    #[specta(type = f64)]
-    pub inventory: Decimal,
-    #[specta(type = f64)]
-    pub min_quantity: Option<Decimal>,
+    #[specta(type = f32)]
+    pub selling_price: Option<f32>,
+    #[specta(type = f32)]
+    pub purchase_price: Option<f32>,
+    #[specta(type = f32)]
+    pub inventory: f32,
+    #[specta(type = f32)]
+    pub min_quantity: Option<f32>,
 }
 
 #[derive(Deserialize, Serialize, Debug, Type)]
@@ -39,20 +38,20 @@ pub struct ProductsResponse {
 pub struct ProductSearch {
     pub label: String,
     pub value: String,
-    #[specta(type = f64)]
-    pub price: Option<Decimal>,
+    #[specta(type = f32)]
+    pub price: Option<f32>,
 }
 
 #[derive(Debug, Serialize, Deserialize, Type)]
 pub struct NewProduct {
     pub name: String,
     pub description: Option<String>,
-    #[specta(type = f64)]
-    pub selling_price: Decimal,
-    #[specta(type = f64)]
-    pub purchase_price: Decimal,
-    #[specta(type = f64)]
-    pub min_quantity: Decimal,
+    #[specta(type = f32)]
+    pub selling_price: f32,
+    #[specta(type = f32)]
+    pub purchase_price: f32,
+    #[specta(type = f32)]
+    pub min_quantity: f32,
     pub image: Option<String>,
 }
 
@@ -61,11 +60,11 @@ pub struct Product {
     pub id: String,
     pub name: String,
     pub description: Option<String>,
-    #[specta(type = f64)]
-    pub selling_price: Decimal,
-    #[specta(type = f64)]
-    pub purchase_price: Decimal,
-    #[specta(type = f64)]
-    pub min_quantity: Decimal,
+    #[specta(type = f32)]
+    pub selling_price: f32,
+    #[specta(type = f32)]
+    pub purchase_price: f32,
+    #[specta(type = f32)]
+    pub min_quantity: f32,
     pub image: Option<String>,
 }
